add rendering of panel in gui.

+++ src/main/java/fr/mattmouss/signs/util/QuadPSPosition.java
@@ -1,18 +1,17 @@
 package fr.mattmouss.signs.util;
 
-import net.minecraft.util.math.Vec2f;
 
 public class QuadPSPosition {
-    private Vec2f position;
-    private float lengthMax;
+    private Vec2i position;
+    private int lengthMax;
     private int maxText;
-    public QuadPSPosition(float x1, float y1, float lengthMax, int maxTextNumber){
-        position = new Vec2f(x1,y1);
+    public QuadPSPosition(int x1, int y1, int lengthMax, int maxTextNumber){
+        position = new Vec2i(x1,y1);
         this.lengthMax = lengthMax;
         this.maxText = maxTextNumber;
     }
 
-    public Vec2f getPosition() {
+    public Vec2i getPosition() {
         return position;
     }
